requirements.txt
colorama==0.4.4
PyYAML==6.0
Telethon==1.26.0
